Added confirmation when deleting appointments. Moved confirmation box logic to the SchedulingManagerForm from DBService for deleting customer.

+++ SoftwareII/Forms/SchedulingManagerForm.cs
@@ -11,6 +11,7 @@ namespace SoftwareII.Forms
     {
         private List<Customer> _allCustomers;
         private List<Appointment> _allAppointments;
+        private List<Appointment> _calendarAppointments;
         public CultureInfo _culture;
 
         public SchedulingManagerForm()
@@ -56,10 +57,15 @@ namespace SoftwareII.Forms
             {
                 for (int i = 0; i < customerDataGrid.SelectedRows.Count; i++)
                 {
-                    Program.DBService.DeleteCustomer((int)customerDataGrid.SelectedRows[i].Cells["customerId"].Value);
+                    var confirmResult = MessageBox.Show("Are you sure you want to delete this customer? It will also delete all appointments and data related to this customer.", "Delete Customer?", MessageBoxButtons.YesNo);
+
+                    if (confirmResult == DialogResult.Yes)
+                    {
+                        Program.DBService.DeleteCustomer((int)customerDataGrid.SelectedRows[i].Cells["customerId"].Value);
+                    }
                 }
             }
-        }
+        } 
 
         public void RefreshData()
         {
@@ -83,9 +89,33 @@ namespace SoftwareII.Forms
             {
                 for (int i = 0; i < appointmentDatagrid.SelectedRows.Count; i++)
                 {
-                    Program.DBService.DeleteAppointment((int)appointmentDatagrid.SelectedRows[i].Cells["appointmentId"].Value);
+
+                    var confirmResult = MessageBox.Show("Are you sure you want to delete this appointment?", "Delete Appointment?", MessageBoxButtons.YesNo);
+
+                    if (confirmResult == DialogResult.Yes)
+                    {
+                        Program.DBService.DeleteAppointment((int)appointmentDatagrid.SelectedRows[i].Cells["appointmentId"].Value);
+                    }
                 }
             }
+        }
+
+        private void weeklyViewRadioButton_CheckedChanged(object sender, EventArgs e)
+        {
+            _calendarAppointments = _allAppointments.FindAll(appointment => appointment.start > DateTime.Now && appointment.start < DateTime.Now.AddDays(8) && appointment.createdBy == Program.AuthService._activeUser);
+            calendarLabel.Text = "Appointment Calendar - This Week (next 7 days)";
+            calendarListView.View = View.Details;
+            calendarListView.Columns.Add("Customer");
+            calendarListView.Columns.Add("Appointment Start");
+            foreach (var appointment in _calendarAppointments)
+            {
+                ListViewItem item = new ListViewItem();
+            }
+        }
+
+        private void monthlyViewRadioButton_CheckedChanged(object sender, EventArgs e)
+        {
+            calendarLabel.Text = "Appointment Calendar - This Month (next 30 days)";
         }
     }
 }
